add tx sync manager and remove handler::query
