+++ core/src/light_protocol/common/mod.rs
@@ -3,20 +3,16 @@
 // See http://www.gnu.org/licenses/
 
 mod ledger_info;
-mod ledger_proof;
 mod peers;
 mod poll;
 mod timeout;
 mod unique_id;
-mod validate;
 
 pub use ledger_info::LedgerInfo;
-pub use ledger_proof::LedgerProof;
-pub use peers::Peers;
+pub use peers::{FullPeerState, LightPeerState, Peers};
 pub use poll::{poll_future, poll_stream};
 pub use timeout::{with_timeout, Timeout};
 pub use unique_id::UniqueId;
-pub use validate::Validate;
 
 use std::cmp;
