+++ core/src/light_protocol/handler/sync/blooms.rs
@@ -12,8 +12,7 @@ use std::{collections::HashMap, sync::Arc, time::Duration};
 use crate::{
     hash::keccak,
     light_protocol::{
-        common::{Peers, UniqueId},
-        handler::FullPeerState,
+        common::{FullPeerState, Peers, UniqueId},
         message::{BloomWithEpoch, GetBlooms},
         Error, ErrorKind,
     },
@@ -26,8 +25,8 @@ use crate::{
 };
 
 use super::{
-    future_item::FutureItem, missing_item::KeyOrdered,
-    sync_manager::SyncManager, witnesses::Witnesses,
+    common::{FutureItem, KeyOrdered, SyncManager},
+    witnesses::Witnesses,
 };
 
 #[derive(Debug)]
@@ -54,7 +53,7 @@ pub struct Blooms {
 }
 
 impl Blooms {
-    pub(super) fn new(
+    pub fn new(
         peers: Arc<Peers<FullPeerState>>, request_id_allocator: Arc<UniqueId>,
         witnesses: Arc<Witnesses>,
     ) -> Self
@@ -94,7 +93,7 @@ impl Blooms {
     }
 
     #[inline]
-    pub(super) fn receive(
+    pub fn receive(
         &self, blooms: impl Iterator<Item = BloomWithEpoch>,
     ) -> Result<(), Error> {
         for BloomWithEpoch { epoch, bloom } in blooms {
@@ -109,7 +108,7 @@ impl Blooms {
     }
 
     #[inline]
-    pub(super) fn clean_up(&self) {
+    pub fn clean_up(&self) {
         let timeout = Duration::from_millis(BLOOM_REQUEST_TIMEOUT_MS);
         let blooms = self.sync_manager.remove_timeout_requests(timeout);
         self.sync_manager.insert_waiting(blooms.into_iter());
@@ -135,7 +134,7 @@ impl Blooms {
     }
 
     #[inline]
-    pub(super) fn sync(&self, io: &dyn NetworkContext) {
+    pub fn sync(&self, io: &dyn NetworkContext) {
         info!("bloom sync statistics: {:?}", self.get_statistics());
 
         self.sync_manager.sync(
