+++ core/src/parameters.rs
@@ -137,6 +137,7 @@ pub mod light {
     pub const BLOCK_TX_REQUEST_TIMEOUT_MS: u64 = 2000;
     pub const STATE_ROOT_REQUEST_TIMEOUT_MS: u64 = 2000;
     pub const STATE_ENTRY_REQUEST_TIMEOUT_MS: u64 = 2000;
+    pub const TX_REQUEST_TIMEOUT_MS: u64 = 2000;
 
     /// Maximum time period we wait for a response for an on-demand query.
     /// After this timeout has been reached, we try another peer or give up.
@@ -154,6 +155,7 @@ pub mod light {
     pub const BLOCK_TX_REQUEST_BATCH_SIZE: usize = 30;
     pub const STATE_ROOT_REQUEST_BATCH_SIZE: usize = 30;
     pub const STATE_ENTRY_REQUEST_BATCH_SIZE: usize = 30;
+    pub const TX_REQUEST_BATCH_SIZE: usize = 30;
 
     /// Maximum number of in-flight items at any given time.
     /// If we reach this limit, we will not request any more.
@@ -164,6 +166,7 @@ pub mod light {
     pub const MAX_BLOCK_TXS_IN_FLIGHT: usize = 100;
     pub const MAX_STATE_ROOTS_IN_FLIGHT: usize = 100;
     pub const MAX_STATE_ENTRIES_IN_FLIGHT: usize = 100;
+    pub const MAX_TXS_IN_FLIGHT: usize = 100;
 
     /// Maximum number of in-flight epoch requests at any given time.
     /// Similar to `MAX_HEADERS_IN_FLIGHT`. However, it is hard to match
