+++ tests/light/sync_test.py
@@ -62,14 +62,14 @@ class LightSyncTest(ConfluxTestFramework):
 
         self.connect_light_node([FULLNODE0, FULLNODE1])
         wait_for_block_count(self.nodes[LIGHTNODE], 1 * block_batch_size + 1)
-        self.log.info(f"Pass 1 - catch up {block_batch_size} blocks from 2 peers")
+        self.log.info(f"Pass 1 - catch up {block_batch_size} blocks from 2 peers\n")
 
         # keep up
         self.generate_blocks(block_batch_size)
         wait_for_block_count(self.nodes[FULLNODE0], 2 * block_batch_size + 1)
         wait_for_block_count(self.nodes[FULLNODE1], 2 * block_batch_size + 1)
         wait_for_block_count(self.nodes[LIGHTNODE], 2 * block_batch_size + 1)
-        self.log.info(f"Pass 2 - keep up with 2 peers for {block_batch_size} blocks")
+        self.log.info(f"Pass 2 - keep up with 2 peers for {block_batch_size} blocks\n")
 
         # catch up again
         self.disconnect_light_node([FULLNODE0, FULLNODE1])
@@ -80,7 +80,7 @@ class LightSyncTest(ConfluxTestFramework):
 
         self.connect_light_node([FULLNODE0])
         wait_for_block_count(self.nodes[LIGHTNODE], 3 * block_batch_size + 1)
-        self.log.info(f"Pass 3 - catch up {block_batch_size} blocks from 1 peer")
+        self.log.info(f"Pass 3 - catch up {block_batch_size} blocks from 1 peer\n")
 
 
 if __name__ == "__main__":
